basic watch mplementation; causes infinite forth and back compilation
