+++ Cargo.toml
@@ -37,3 +37,4 @@ glob = "0.3"
 clap = "2.33"
 blake3 = "0.3"
 once_cell = "1.5"
+notify = "4.0"
